requirements.txt
Flask==2.2.3
google-cloud-vision==3.3.1
google-cloud-translate==3.12.1
openai==0.27.0
pillow==9.2.0
pytesseract==0.3.9
requests==2.28.1
opencv-python==4.8.0.76
numpy==1.23.5
